+++ src/audiovideocontroller/DefaultAudioVideoController.ts
@@ -46,6 +46,7 @@ import ReceiveAudioInputTask from '../task/ReceiveAudioInputTask';
 import ReceiveTURNCredentialsTask from '../task/ReceiveTURNCredentialsTask';
 import ReceiveVideoInputTask from '../task/ReceiveVideoInputTask';
 import ReceiveVideoStreamIndexTask from '../task/ReceiveVideoStreamIndexTask';
+import SendAndReceiveDataMessagesTask from '../task/SendAndReceiveDataMessagesTask';
 import SerialGroupTask from '../task/SerialGroupTask';
 import SetLocalDescriptionTask from '../task/SetLocalDescriptionTask';
 import SetRemoteDescriptionTask from '../task/SetRemoteDescriptionTask';
@@ -271,6 +272,7 @@ export default class DefaultAudioVideoController implements AudioVideoController
               new SerialGroupTask(this.logger, 'Signaling', [
                 new OpenSignalingConnectionTask(this.meetingSessionContext),
                 new ListenForVolumeIndicatorsTask(this.meetingSessionContext),
+                new SendAndReceiveDataMessagesTask(this.meetingSessionContext),
                 new JoinAndReceiveIndexTask(this.meetingSessionContext),
                 // TODO: ensure index handler does not race with incoming index update
                 new ReceiveVideoStreamIndexTask(this.meetingSessionContext),
@@ -638,7 +640,6 @@ export default class DefaultAudioVideoController implements AudioVideoController
 
   async handleHasBandwidthPriority(hasBandwidthPriority: boolean): Promise<void> {
     if (this.meetingSessionContext && this.meetingSessionContext.videoUplinkBandwidthPolicy) {
-      this.logger.info(`video send has bandwidth priority: ${hasBandwidthPriority}`);
       const oldMaxBandwidth = this.meetingSessionContext.videoUplinkBandwidthPolicy.maxBandwidthKbps();
       this.meetingSessionContext.videoUplinkBandwidthPolicy.setHasBandwidthPriority(
         hasBandwidthPriority
@@ -646,7 +647,7 @@ export default class DefaultAudioVideoController implements AudioVideoController
       const newMaxBandwidth = this.meetingSessionContext.videoUplinkBandwidthPolicy.maxBandwidthKbps();
       if (oldMaxBandwidth !== newMaxBandwidth) {
         this.logger.info(
-          `video send bandwidth max has changed from ${oldMaxBandwidth} kbps to ${newMaxBandwidth} kbps`
+          `video send bandwidth priority ${hasBandwidthPriority} max has changed from ${oldMaxBandwidth} kbps to ${newMaxBandwidth} kbps`
         );
         await this.enforceBandwidthLimitationForSender(newMaxBandwidth);
       }