+++ test/dommock/DOMMockBuilder.ts
@@ -500,6 +500,9 @@ export default class DOMMockBuilder {
               mediaStreamTrack.kind =
                 /^m=video/gm.exec(description.sdp) !== null ? 'video' : 'audio';
               addTrackEvent.track = mediaStreamTrack;
+              if (mockBehavior.hasInactiveTransceiver) {
+                addTrackEvent.transceiver = { currentDirection: 'inactive' };
+              }
               const mediaStreamMaker: typeof GlobalAny.MediaStream = GlobalAny.MediaStream;
               const mediaStream = new mediaStreamMaker();
               mediaStream.id = mockBehavior.setRemoteDescriptionStreamId;
@@ -753,11 +756,7 @@ export default class DOMMockBuilder {
         return mockBehavior.FakeTURNCredentialsBody;
       }
       get status(): number {
-        if (mockBehavior.responseSuccess) {
-          return 200;
-        } else {
-          return 500;
-        }
+        return mockBehavior.responseStatusCode;
       }
     };